requirements.txt
pocketflow>=0.0.1
requests>=2.28.0
beautifulsoup4>=4.11.0
youtube-transcript-api>=0.6.0
openai>=1.0.0
pyyaml>=6.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0